renamed event_listener.py to input_trigger.py
